・連射武器の射撃間隔が1フレーム遅れている不具合修正 ・弾を発射するまでの時間を設定できるようにした ・ショットガン実装。ついでにどの武器でも一度に撃つ弾の数と弾の拡散範囲を設定できるようにした ・バースト銃の実装。ついでにどの武器でもバースト間隔、バースト後のクールタイム、何連射かを決められるようにした

+++ class/GameObjects/Weapon/Components/Magazine.cpp
@@ -1,21 +1,21 @@
 #include "Magazine.h"
 #include <algorithm>
 
-Magazine::Magazine(const float& bulletNum) {
+Magazine::Magazine(const int& bulletNum) {
 	// 残弾を代入
 	Init(bulletNum);
 }
 
-void Magazine::Init(const float& bulletNum) {
+void Magazine::Init(const int& bulletNum) {
 	// 残弾を代入
 	bulletNum_ = bulletNum;
 }
 
 void Magazine::Update() {
 	// 弾数が0を下回らないようにする
-	bulletNum_ = std::min<float>(bulletNum_, 0.0f);
+	bulletNum_ = std::min<int>(bulletNum_, 0.0f);
 }
 
 void Magazine::DebugGui() {
-	ImGui::DragFloat("BulletCount", &bulletNum_);
+	ImGui::DragInt("BulletCount", &bulletNum_);
 }
